Move shared components to src/shared
