+++ packages/site-client/src/shared/wco-footer.ts
